Made more Py3K- and PEP8-friendly.
Python 3's `map` function returns an iterator, not a list.  This code
was operating under the old, Python 2.x assumption.  There's little
reason *not* to do the explicit conversion, as it still functions the
same way in my IDLE 2.7 tests, with the benefit of Python 3 not
complaining about the fact that the `'map' object has no attribute
'append'`.

Lastly, before submission, `autopep8 --in-place` was run on the code.

Signed-off-by: Jonathan Sifuentes <jayands2k11@ymail.com>